requirements.txt
wheel==0.37.0
aniso8601==9.0.1
cachelib==0.4.1
click==8.0.3
colorama==0.3.9
Deprecated==1.2.13
dominate==2.6.0
et-xmlfile==1.1.0
Flask==2.0.2
flask-nav==0.6
Flask-Session==0.4.0
Flask-WTF==1.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
lxml==4.7.1
MarkupSafe==2.0.1
nltk==3.6.5
openpyxl==3.0.9
PyJWT==2.3.0
python-docx==0.8.11
python-dotenv==0.19.2
pytz==2021.3
redis==4.0.2
regex==2021.11.10
six==1.16.0
tqdm==4.62.3
visitor==0.1.3
Werkzeug==2.0.2
wordnet==0.0.1b2
wrapt==1.13.3
WTForms==3.0.0

